+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Improved command loading performance by implementing lazy loading.
 - [BREAKING] Changed `firestore:backups:list --json` to return a `listBackupsResponse` object instead of a raw array of backups.
 - [BREAKING] Removed support for '.bolt' rules files.
 - [BREAKING] Removed support for running emulators with Java versions prior to 21.